requirements.txt
streamlit==1.34.0
numpy==1.23.5
Pillow==10.0.0
tensorflow-cpu==2.11.0
gdown==5.2.0
requests
